refactor(benchmark): split RouteSelector into Dot and InlinePicker files
